requirements.txt
numpy
pandas
datasets
transformers
torch
torchvision
torchmetrics
einops
peft
pillow
accelerate
bitsandbytes
sentence-transformers